requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
pytz==2024.1
supabase>=2.0.0
